Rewrite class SpeechRecognition to avoid bugs
Previous implementation didn't work properly, if you went to ActivitySettings and fast clicked speechRecogntion On and Off.
Given the issue I implement a HelperQueue that handles which runnable (Startup Speech Recognizer, Shutdown Speech Recognizer) starts.
It checks if there is currently any runnable running. If not input runnable is started. If there is runnable running, it checks if runnable is different than
running runnable. If input runnable is different than runing runnable it goes to queue and it is triggered when this function closes (queue poll). But in Start Speech Recognizer there is AsyncClass,
so we need to watch where to trigger queue poll. If input runnable is the same that current, which is running, we delete queue.