Update GraphActivity with proper values from sensors for representation
Move low-pass filtering to MainSensorManager. This also implies for LightActivity and of course for all sensors. Values for most of sensors in GraphActivity are now normalized. This does not imply on gyroscope and linear acceleration. In this scenario we divide values with maximum value from sensors and then multiply with some number to get proper visual representation.